Added classes, Map, Movement
Used modified code from the 2d map project to draw a simple map from a mapper class. Made a player class that can move using a short loop called at the end of the main function of the Main class. Made an enemy class, it is unused at the moment.